# docs/openapi.yaml
openapi: 3.0.1
info:
  title: Arborist
  version: 1.1.1
  description: authorization microservice to handle RBAC based on configured policies
  license:
    name: 'Apache 2.0'
    url: 'https://github.com/uc-cdis/arborist'
...
paths:
  /auth/request:
    post:
      tags:
        - auth
      description: >-
        Ask arborist to check authorization on a user request. A client of the
        Gen3 auth stack sends the user's JWT to this endpoint, along with the
        information about the user's specific request. The terminology used to
        describe user requests is flexible and arborist does not "nail it down";
        however our usage so far is to use the permission model to capture the
        basic set of operations one might expect in RBAC (read/write, etc.)
        combined with the client service, and the path for the resource the user
        is trying to access.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRequestBody'
      responses:
        200:
          description: >-
            Arborist successfully returned an authorization decision. NOTE
            that a 200 status DOES NOT indicate authorization, only that the
            input was valid. The `"auth"` field in the response indicates
            authorization.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthRequestResponse'
        400:
          description: >-
            The input was somehow invalid; for example, a given resource does
            not exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserError'
  /auth/resources:
    post:
      tags:
        - auth
      description: >-
        Given a user token, return a list of resources which are accessible to
        the user (using any action) according to the policies listed in the
        token.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthResourcesRequestBody'
      responses:
        200:
          description: >-
            Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResourcesResponse'
        400:
          description: invalid input (missing fields or fields have incorrect types)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserError'
        401:
          description: >-
            Token failed to validate (authentication error)
  /auth/proxy:
    get:
      tags:
        - auth
      description: >-
        The proxy endpoint works differently from the other auth endpoints.
        Whereas the other endpoints will return a 200 code as long as the
        request was understood (and allowed), whether or not the auth request
        should be appproved or denied, this endpoint will return a code and
        error JSON as if to the end user making the request that is checked,
        whether or not the request is valid.
      parameters:
        - in: header
          name: Authorization
          schema:
            type: string
          required: true
        - in: query
          name: resource
          schema:
            type: string
          required: true
        - in: query
          name: service
          schema:
            type: string
          required: true
        - in: query
          name: method
          schema:
            type: string
          required: true
      responses:
        200:
          description: request is authorized
        401:
          description: authentication failed (the token was invalid)
        403:
          description: authorization failed (user denied access)
...
components:
  schemas:
    AuthRequestBody:
      type: object
      properties:
        user:
          type: object
          properties:
            jwt:
              type: string
              description: >-
                A JWT belonging to a user, as defined by RFC 7519. It must contain a
                field `context.user.policies` listing the policies granted to the user.
              example: 'eyJhbGciOiJFUzI1NiIsImtpZCI6IjE2In0[...]'
        request:
          type: object
          properties:
            resource:
              type: string
              example: '/programs/DEV/projects/test'
            action:
              type: object
              properties:
                service:
                  type: string
                  example: 'fence'
                method:
                  type: string
                  example: 'read-storage'
              required:
                - service
                  method
          required:
            - token
    AuthRequestResponse:
      type: object
      properties:
        auth:
          type: boolean
    AuthResourcesRequestBody:
      type: object
      properties:
        user:
          type: object
          properties:
            token:
              type: string
              example: 'eyJhbGciOiJFUzI1NiIsImtpZCI6IjE2In0[...]'
          required:
            - token
      required:
        - user
    AuthResourcesResponse:
      type: object
      properties:
        resources:
          type: array
          items:
            type: string
          example: ['/programs/DEV/projects/test', '/programs/foo/projects/bar']
    UserError:
      type: object
      properties:
        error:
          type: object
          properties:
            message:
              type: string
              description: an explanation of the error that occurred
            code:
              type: integer
              description: the HTTP error code
      example:
        error:
          message: "input resource is missing the following required fields: ..."
          code: 400
...